added documentation and removed src folder.
